feat(AbrigoDTO): add dto to show all data for Abrigo
