+++ api/src/main/java/br/com/adopet/api/dto/abrigo/AbrigoDTO.java
@@ -1,18 +1,19 @@
-package br.com.adopet.api.dto;
+package br.com.adopet.api.dto.abrigo;
 
+import br.com.adopet.api.domain.model.Pet;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class TutorDTO {
+public class AbrigoDTO {
 
     private Long id;
     private String nome;
-    private String telefone;
-    private String cidade;
-    private String sobre;
+    private List<Pet> pets;
 
 }
