+++ api/src/main/java/br/com/adopet/api/dto/tutor/DadosAtualizarTutor.java
@@ -1,4 +1,4 @@
-package br.com.adopet.api.dto;
+package br.com.adopet.api.dto.tutor;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -9,16 +9,15 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class DadosCadastroTutor {
+public class DadosAtualizarTutor {
 
-    @NotBlank
     private String nome;
-    @NotBlank
+
     @Pattern(regexp = "\\(\\d{2}\\)\\d{5}-\\d{4}")
     private String telefone;
-    @NotBlank
+
     private String cidade;
-    @NotBlank
+
     private String sobre;
 
 }
