+++ api/src/main/java/br/com/adopet/api/dto/tutor/TutorDTO.java
@@ -1,0 +1,18 @@
+package br.com.adopet.api.dto.tutor;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class TutorDTO {
+
+    private Long id;
+    private String nome;
+    private String telefone;
+    private String cidade;
+    private String sobre;
+
+}
